+++ Sources/Core/State/LoadedState.swift
@@ -75,7 +75,7 @@ struct LoadedState: PlayerState {
     }
 
     func seek(position: Double) {
-        let time = CMTime(seconds: position, preferredTimescale: context.config.preferedTimeScale)
+        let time = CMTime(seconds: position, preferredTimescale: context.config.preferredTimescale)
         context.player.seek(to: time) { [context] completed in
             guard completed else { return }
             context.delegate?.playerContext(didCurrentTimeChange: context.currentTime)